DAOS-9690 test: Add support to use label in tests for daos container … (#7958)
The goal is to convert the tests to use label to create and destroy
container. This makes it a complex PR, so I broke it down to 3 parts.
This is the first part that does the following:

1. Move LabelGenerator from test_utils_pool.py to test_utils_base.py
because both TestPool and TestContainer use it.
2. Add self.label in daos_utils_base.py -> class CreateSubCommand
to add --label support.
3. Add label parameter in daos_utils.py -> container_create() and
pass it in to _get_json_result()
