+++ src/tests/ftest/nvme/pool_exclude.py
@@ -9,7 +9,8 @@ import random
 import threading
 import re
 
-from test_utils_pool import TestPool, LabelGenerator
+from test_utils_pool import TestPool
+from test_utils_base import LabelGenerator
 from osa_utils import OSAUtils
 from write_host_file import write_host_file
