+++ src/tests/ftest/osa/offline_extend.py
@@ -6,7 +6,8 @@
 """
 from osa_utils import OSAUtils
 from daos_utils import DaosCommand
-from test_utils_pool import TestPool, LabelGenerator
+from test_utils_pool import TestPool
+from test_utils_base import LabelGenerator
 from dmg_utils import check_system_query_status
 from apricot import skipForTicket
