+++ src/tests/ftest/osa/online_reintegration.py
@@ -8,7 +8,8 @@ import time
 import random
 import threading
 
-from test_utils_pool import TestPool, LabelGenerator
+from test_utils_pool import TestPool
+from test_utils_base import LabelGenerator
 from write_host_file import write_host_file
 from daos_racer_utils import DaosRacerCommand
 from osa_utils import OSAUtils
