+++ src/tests/ftest/util/daos_utils.py
@@ -61,7 +61,7 @@ class DaosCommand(DaosCommandBase):
 
     def container_create(self, pool, sys_name=None, cont=None,
                          path=None, cont_type=None, oclass=None,
-                         chunk_size=None, properties=None, acl_file=None):
+                         chunk_size=None, properties=None, acl_file=None, label=None):
         # pylint: disable=too-many-arguments
         """Create a container.
 
@@ -80,6 +80,7 @@ class DaosCommand(DaosCommandBase):
             properties (str, optional): String of comma-separated <name>:<value>
                 pairs defining the container properties. Defaults to None
             acl_file (str, optional): ACL file. Defaults to None.
+            label (str, optional): Container label. Defaults to None.
 
         Returns:
             dict: the daos json command output converted to a python dictionary
@@ -91,7 +92,7 @@ class DaosCommand(DaosCommandBase):
         return self._get_json_result(
             ("container", "create"), pool=pool, sys_name=sys_name,
             cont=cont, path=path, type=cont_type, oclass=oclass,
-            chunk_size=chunk_size, properties=properties, acl_file=acl_file)
+            chunk_size=chunk_size, properties=properties, acl_file=acl_file, label=label)
 
     def container_clone(self, src, dst):
         """Clone a container to a new container.
