+++ src/tests/ftest/util/daos_utils_base.py
@@ -278,6 +278,8 @@ class DaosCommandBase(CommandWithSubCommand):
                 #    -c, --cont=<UUID>
                 #           container UUID (optional)
                 self.cont = FormattedParameter("--cont={}")
+                #     -l, --label=<container label>
+                self.label = FormattedParameter("--label={}")
 
         class CreateSnapSubCommand(CommonContainerSubCommand):
             """Defines an object for the daos container create-snap command."""
