+++ src/tests/ftest/util/test_utils_pool.py
@@ -9,7 +9,7 @@ from time import sleep, time
 import ctypes
 import json
 
-from test_utils_base import TestDaosApiBase
+from test_utils_base import TestDaosApiBase, LabelGenerator
 from avocado import fail_on
 from command_utils import BasicParameter, CommandFailure
 from pydaos.raw import (DaosApiError, DaosPool, c_uuid_to_str, daos_cref)
@@ -941,32 +941,3 @@ class TestPool(TestDaosApiBase):
                 pool=self.identifier, acl_file=self.acl_file.value)
         else:
             self.log.error("self.acl_file isn't defined!")
-
-
-class LabelGenerator():
-    # pylint: disable=too-few-public-methods
-    """Generates label used for pool."""
-
-    def __init__(self, value=1):
-        """Constructor.
-
-        Args:
-            value (int): Number that's attached after the base_label.
-        """
-        self.value = value
-
-    def get_label(self, base_label):
-        """Create a label by adding number after the given base_label.
-
-        Args:
-            base_label (str): Label prefix. Don't include space.
-
-        Returns:
-            str: Created label.
-
-        """
-        label = base_label
-        if label is not None:
-            label = "_".join([base_label, str(self.value)])
-            self.value += 1
-        return label
